requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
django==5.1.7
kombu==5.5.0
prompt-toolkit==3.0.50
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
